pull all the streambot code into index.js
